backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
Pillow==9.5.0
matplotlib==3.7.5
numpy==1.24.3
python-dotenv==1.0.0
PyJWT==2.8.0
celery==5.3.4
redis==5.0.1
sqlalchemy==1.4.53
psycopg2-binary==2.9.7
APScheduler==3.10.4
opencv-python-headless==4.8.1.78
boto3==1.34.0
botocore==1.34.0